# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API",
    "version": "1.0.0",
    "description": "API documentation for my Express application"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/teacher/courses": {
      "get": {
        "summary": "Retrieve all courses",
        "responses": {
          "200": {
            "description": "A list of courses"
          }
        }
      },
      "post": {
        "summary": "Create a new course",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully"
          }
        }
      }
    },
...
    "/teacher/assignments": {
      "post": {
        "summary": "Create a new assignment",
        "tags": [
          "Teacher"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assignment created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "summary": "View all assignments",
        "tags": [
          "Teacher"
        ],
        "responses": {
          "200": {
            "description": "Successful response with a list of assignments"
          }
        }
      }
    },
...
    "/teacher/upload": {
      "get": {
        "summary": "Render the file upload form",
        "responses": {
          "200": {
            "description": "Render upload form"
          }
        }
      },
      "post": {
        "summary": "Upload a file",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully"
          }
        }
      }
    },
...
  },
...
}